+++ src/main/java/com/project/demo/rest/product/ProductRestController.java
@@ -1,15 +1,24 @@
 package com.project.demo.rest.product;
 import com.project.demo.logic.entity.category.Category;
 import com.project.demo.logic.entity.category.CategoryRepository;
+import com.project.demo.logic.entity.http.GlobalResponseHandler;
+import com.project.demo.logic.entity.http.Meta;
 import com.project.demo.logic.entity.product.Product;
 import com.project.demo.logic.entity.product.ProductRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
+import javax.swing.text.html.Option;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/products")
@@ -31,8 +40,21 @@ public class ProductRestController {
 
     @GetMapping
     @PreAuthorize("isAuthenticated()")
-    public List<Product> getAllProducts(){
-        return productRepository.findAll();
+    public ResponseEntity<?> getAllProducts(
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int size,
+            HttpServletRequest request
+    ){
+        Pageable pageable = PageRequest.of(page-1, size);
+        Page<Product> productsPage = productRepository.findAll(pageable);
+        Meta meta = new Meta(request.getMethod(), request.getRequestURL().toString());
+        meta.setTotalPages(productsPage.getTotalPages());
+        meta.setTotalElements(productsPage.getTotalElements());
+        meta.setPageNumber(productsPage.getNumber() + 1);
+        meta.setPageSize(productsPage.getSize());
+
+        return new GlobalResponseHandler().handleResponse("Products retrieved successfully",
+                productsPage.getContent(), HttpStatus.OK, meta);
     }
 
     @GetMapping("/{id}")
@@ -44,31 +66,39 @@ public class ProductRestController {
 
     @PutMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN')")
-    public Product updateProduct(@PathVariable Long id, @RequestBody Product product) {
-        return productRepository.findById(id)
-                .map(existingProduct-> {
-                    existingProduct.setName(product.getName());
-                    existingProduct.setDescription(product.getDescription());
-                    existingProduct.setPrice(product.getPrice());
-                    existingProduct.setStock(product.getStock());
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody Product product, HttpServletRequest request) {
+        Optional<Product> foundProduct = productRepository.findById(id);
 
-                    Category category = categoryRepository.findById(product.getCategory().getId())
-                            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria no encontrada"));
-                    existingProduct.setCategory(category);
-                    return productRepository.save(existingProduct);
-                })
-                .orElseGet(() -> {
-                    product.setId(id);
-                    Category category = categoryRepository.findById(product.getCategory().getId())
-                            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Categoria no encontrada"));
-                    product.setCategory(category);
-                    return productRepository.save(product);
-                });
+        if(foundProduct.isPresent()) {
+            Optional<Category> foundCategory = categoryRepository.findById(product.getCategory().getId());
+            product.setId(id);
+            product.setName(product.getName());
+            product.setDescription(product.getDescription());
+            product.setPrice(product.getPrice());
+            product.setStock(product.getStock());
+            product.setCategory(foundCategory.get());
+            productRepository.save(product);
+
+            return new GlobalResponseHandler().handleResponse("Product updated successfully",
+                    product, HttpStatus.OK, request);
+        } else {
+            return new GlobalResponseHandler().handleResponse("Product id " + id + " not found"  ,
+                    HttpStatus.NOT_FOUND, request);
+        }
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN')")
-    public void deleteProduct (@PathVariable Long id) {
-        productRepository.deleteById(id);
+    public ResponseEntity<?> deleteProduct (@PathVariable Long id, HttpServletRequest request) {
+        Optional<Product> foundProduct = productRepository.findById(id);
+
+        if(foundProduct.isPresent()) {
+            productRepository.deleteById(id);
+            return new GlobalResponseHandler().handleResponse("Product deleted successfully",
+                    foundProduct.get(), HttpStatus.OK, request);
+        }else {
+            return new GlobalResponseHandler().handleResponse("Product id " + id + " not found"  ,
+                    HttpStatus.NOT_FOUND, request);
+        }
     }
 }
